+++ RuriLib.Parallelization/ThreadBasedParallelizer.cs
@@ -18,7 +18,7 @@ namespace RuriLib.Parallelization
         #region Constructors
         /// <inheritdoc/>
         public ThreadBasedParallelizer(IEnumerable<TInput> workItems, Func<TInput, CancellationToken, Task<TOutput>> workFunction,
-            int degreeOfParallelism, int totalAmount, int skip = 0) : base(workItems, workFunction, degreeOfParallelism, totalAmount, skip)
+            int degreeOfParallelism, long totalAmount, int skip = 0) : base(workItems, workFunction, degreeOfParallelism, totalAmount, skip)
         {
 
         }
