+++ RuriLib/Models/Data/DataPool.cs
@@ -8,7 +8,7 @@ namespace RuriLib.Models.Data
         public IEnumerable<string> DataList { get; protected set; }
 
         /// <summary>The total number of lines.</summary>
-        public int Size { get; protected set; }
+        public long Size { get; protected set; }
 
         /// <summary>The wordlist type for data slicing.</summary>
         public string WordlistType { get; protected set; }
